Balanced Brackets Problem
