+++ cppsolutions/cppsolutions.vcxproj
@@ -147,6 +147,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="BalancedBrackets.cpp" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="DynamicArray.cpp" />
     <ClCompile Include="SparseArray.cpp" />
