+++ cppsolutions/cppsolutions.vcxproj.filters
@@ -30,5 +30,8 @@
     <ClCompile Include="WaterBlockProblem.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="BalancedBrackets.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
 </Project>
